+++ examples/resources/fmc_range_overrides/import.sh
@@ -1,0 +1,2 @@
+# <domain> is optional. If not provided, `Global` is used implicitly and resource's `domain` attribute is not set.
+terraform import fmc_range_overrides.example "<domain>,<id>"